OM-255 | Configure database seeding during deployment
The CI build executes on_deploy.sh during deployment to initialize the db.
I added a command for running the data seeding script to get the database
set up with users, profiles etc. I also removed the scripts for creating
municipalities etc, because we're not currently doing anything with those.