OM-304 | Remove unnecessary fields from the profile API and model
Removed Profile.preferences from the model. It wasn't even exposed by the API.

Removed concepts_of_interest and divisions_of_interest from the API. I left
the model as-is and also preserved the rest of the functionality regarding
these fields. These will probably be re-implemented in the future, since
the interests of a user are valuable information.