OM-359 | Fix the email sending for confirmed youth profile
The "to" email address was set wrong. The value came from get_default_email
which is actually a function. I fixed this so that the default email is
fetched correctly. Added handling for the situation where there is no
primary email address. In this case a new exception is raised.