Make YouthProfile.membership_number a DB field
A DB field for membership number is added to YouthProfile.
A post save signal is used for generating a membership number
for a new youth profile or a profile missing a membership
number i.e. re-saving the profile doesn't overwrite an
existing the membership number.

DB field is used instead of dynamically calculating the membership
number out of the pk of the YouthProfile instance.

Refs OM-698