Allow re-running generate_subscription_types
Running the function will create subscription types and
categories only in case they don't already exist. Nothing
will get deleted and categories or types are not updated.
In this case updating is considered to mean that translations
are not changed by the function once the objects already exist.
Running the function will not delete existing subscriptions.

Refs OM-753