OM-819 | Change the session cookie setting to use secure cookies
Changed the settings so that the server is now using secure cookies if DEBUG
is not enabled so that it's used only in those environments where we have SSL
properly enabled. This issue was brought up by the security auditors and is
also recommended to fix in the Django documentation.