Mark count fields on ProfileConnection as required
This fields are non-nullable by themselves, since if there are
no errors, they will always have some integer value, and if there
are errors either the entire connection will be 'null' or single
nodes on that connection.

Venepaikka also has these as non-nullable and in order to federate
the schemas between our two services all the fields should have
the same scalar types.