Add GraphQL API to create a temporary profile read access token
An UUID is generated as the token. The token has an expiration time and
in this commit the validity time is fixed to two days. The API returns
both the token and the expiration timepoint.

The access token is created for the profile of the currently logged in
user.

The token can't yet be used for anything, there's no API for that.

HP-59