HP-306 Dedicated input types for createProfile and updateProfile mutations
The same input type doesn't fit for all the different kinds of uses.
Hence it's best to have a dedicated input type for each kind of mutation
so that they can evolve their own ways.

This is visible already in this commit as some fields are marked as
deprecated in the CreateProfileInput type. Although GraphQL has a
dedicated deprecation mechanism, it's very limited on where it can be
used. Especially it can't be used for input type fields and that's why
the deprecation has been indicated with just free text documentation
here.