HP-306 Use real UUIDs in faked Users' uuid fields
The Faker's uuid4 generator is weird in the sense that by default it
returns a `str` and not an `UUID`. It is possible to provide the
generator with a `cast_to` argument which changes the type.

In Faker 4.1.0 a new feature was added where it's possible to set the
`cast_to` argument to `None` and get a pure `UUID` object. Before 4.1.0
it was possible to achieve the same by providing an "identity lambda"
(lambda x: x). The identity lambda looks a bit weird so in order to use
the newer style Faker needed a version update.