Cleanup database after migration test
The only migration test that currently exists,
test_fix_primary_email_migration, didn't clean up after itself. It rolls
back in migrations during the test and left the database schema there
after running.

This probably worked before since Pytest seems to order the test
functions so that tests needing transactional database (which includes
the test_fix_primary_email_migration test) are run last - or at least
last within the set of currently existing tests. Only a few more tests
are run after test_fix_primary_email_migration and they don't seem to
depend that much on the database schema and happen to succeed.

But trying to add a new test that uses transactional database and
depends on the database schema to be up-to-date failed. That's why this
needs fixing now.

A library exists for writing migration tests easier:
https://github.com/wemake-services/django-test-migrations.
I chose not to introduce that into the project just yet, as currently it
seems to not bring much value on top of the already existing setup. But
I did get the idea how to reset the database schema from that code base:
by using the 'migrate' management command.