Remove allauth dependency
The social login system was not properly configured and not in use. The
allauth system of django-helusers is deprecated anyway, so if the social
login feature is ever needed, a current way should be used.

The database tables created by allauth are NOT removed by this commit.
Here are the table names for reference:
- account_emailconfirmation
- account_emailaddress
- socialaccount_socialtoken
- socialaccount_socialapp_sites
- socialaccount_socialapp
- socialaccount_socialaccount