Remove stale allauth tables and contenttypes
In commit 5232d4f213 the dependency to django-allauth was removed. But
that commit didn't clean allauth stuff away from the database. This
commit does that.

A migration file handles the cleaning. The migration file is added to
the open_city_profile app - which wasn't an app before, but now it is.
That's why an AppConfig was added to it and it is now also found from
the INSTALLED_APPS setting. Only apps can contain migrations, projects
can't.