Add integration test for testing authentication
Checks what the GraphQL API responds when trying to execute an operation
that requires authentication but no access token is presented.

This test exercises the GraphQL API. But unlike all the existing GraphQL
tests this doesn't use the Graphene test client which takes a shortcut
into the GraphQL query processing. Instead this starts a live server and
makes a real HTTP request. This way the whole stack can be tested,
including authentication processing.