HP-416 Replace graphql_jwt's authentication middleware with our own
This commit makes it possible to opt-in to use a new authenticator
service from django-helusers by setting the
USE_HELUSERS_REQUEST_JWT_AUTH env var to True. By default everything is
as before. Later on the older authentcation system (based on
graphql_jwt) and the feature flag can be removed.

The reasoning for the need of a new authenticator follows.

Django's authentication system is tightly coupled with HTTP (naturally,
since it's an HTTP server framework). That means that Django assumes
that authentication happens once per HTTP request, perhaps early in the
request handling process and if authentication fails, the HTTP response
code is for example 403 Forbidden. With GraphQL, pretty much none of
these assumptions are good.

HTTP is not the only transfer protocol that can be used with GraphQL.
Consequently authentication/authorization doesn't use HTTP status codes
either. Instead, every GraphQL response via HTTP always has status code
200. Any authentication, authorization and other error cases are
reported within the response body.

It's also possible that the caller's credentials give access to only
some part of the requested data. In this case the response can contain
the part of data the caller has access to, and also error(s) telling
what requested data wasn't returned and the reasons why. For this reason
with GraphQL authentication can happen once but authorization can happen
multiple times. Django doens't really give any help for this kind of
scenario as it tightly couples authentication with authorization.

This all means that Django's authentication system isn't really suitable
to be used with GraphQL.

Graphene's middleware system [1] can be used to handle the authorization
problem. The middlewares are called for each field that is to be solved
for the GraphQL response and this is an appropriate point to determine
authorization with suitable granularity.

The graphql_jwt package actually uses Graphene's middleware system. But
its design is flawed: its middleware still uses Django's authentication
system [2] and as was stated above, it's not suitable for this job.

Solving this is quite easy though. In the beginning of each request
handling authentication is done, meaning a User or an error is
generated. The authentication result is stored into the request object.

A Graphene middleware handles authorization for each field, using the
authentication result stored into the request. That's exactly how the
new system in this commit works.

[1] https://docs.graphene-python.org/en/latest/execution/middleware/
[2] https://github.com/flavors/django-graphql-jwt/blob/28e4f9749bac839d327914cfdda2ea3bb77bd775/graphql_jwt/middleware.py#L73