Change some GraphQL tests to use the query+variables style
This concerns the createOrUpdateProfileWithVerifiedPersonalInformation
mutation tests. The tests are exactly same after this commit, only the
querying style changes.

The reason for this change is that it was found that Graphene handles
the query a bit differently depending on whether the input values are
provided within the query or as separate variables. The query+variables
style is the one that clients usually use (3rd party libraries seem to
use that style extensively).