Replace Profile resolvers with DataLoaders
The Node types had to be updated to support the DataLoaders.
DjangoFilterConnectionField doesn't work with DLs since the DL returns a list rather than a QuerySet, so the filtering
can't be done by default by the FilterSet.

In case some filtering would be required, it has to be done manually on the resolver and then pass the IDs of the
filtered items as parameters to the dataloader.