HP-292 Include basic verified personal information into myProfile query
The first level data can be queried. The second level data (addresses)
can't — yet.

Introduced a new type to be returned from myProfile, as ProfileNode (the
previous return type) is used in many places and we don't want to add
the verified personal information everywhere.

The 'profile' field in SubscriptionNode needed to be defined now
explicitly because automagic schema generation would otherwise choose
the VerifiedPersonalInformationNode type but it needs to be ProfileNode
as it was previously. The lambda in the field definition is required in
order to circumvent a circular dependency between subscription schema
and profiles schema.