HP-398 Enable filtering profiles with ids
The "profiles" GraphQL query got a new filtering option: `id`. A list of
UUIDs can be provided to select only the wanted profiles.

Since a "list of UUIDs" is a type not directly supported by Graphene
and/or django_filters, needed to implement some small classes to get the
type of the new filtering argument correct.