HP-178 Resolve acting user correctly in audit log
The previous version which tried to put the authenticated user (a.k.a.
actor) into the (thread local) global state didn't work in all cases.
Especially with the GraphQL API the code that tried to set the user was
ran before authentication took place. As a result, it always inserted
the AnonymousUser into the global state.

Now the same middleware that used to set the user, puts the whole
request object into the global state. Later when authentication happens
the user gets inserted into the request. Then when audit logging happens
(always after authentication) the user can be extracted from the request
that is found from the global state.