HP-556 Change audit logging to "buffered" type
The previous implementation potentially caused multiple extra SQL
queries to be made. The Profile model's post_init signal was attached
directly to the creation of the audit log event. The documentation for
that signal [1] contains a warning:

"For performance reasons, you shouldn’t perform queries in receivers of
 pre_init or post_init signals because they would be executed for each
 instance returned during queryset iteration."

But that’s exactly what the audit logging system did. It requested
profile.user and that potentially needed to query the database for the
User.

The audit logging system is changed in a way that post_init signal only
adds the model instance to be audit logged. The actual audit log event
creation happens later. Currently a request middleware does this audit
log "flushing".

[1] https://docs.djangoproject.com/en/3.1/ref/signals/#post-init