HP-536 Change name fields to unencrypted
Changes the first_name and last_name fields of
VerifiedPersonalInformation from encrypted fields to normal CharFields.

After this migration the names of the fields stay the same and existing
data is preserved. The migration needs to be done in two steps (i.e. in
two transactions) because reading data from a column and removing that
column can't happen within the same transaction.