HP-91 Enable deleting the related User when deleting a Profile
The Profile deletion confirmation screen contains a checkbox that can be
used to control if the related User should get deleted as well.

Extending the confirmation screen is a bit cumbersome. In this commit
the content block part of the original template is copied and a few
lines of custom HTML is inserted into it. The inserted part is indicated
with comments in the template.