HP-575 Determine requester's Service by authentication
The Service is determined from authentication data, if it's available.

Some GraphQL operations have a `serviceType` argument or input field
meant to serve this same purpose. Changed the `serviceType` argument in
the `profiles` query as nullable and marked it as deprecated to make
testing of this new feature possible.

When the `serviceType` argument is provided, it overrides the Service
determined by authentication. The reason is that the requester can in
theory provide different `serviceTypes` for different GraphQL
operations, within a single request, something like this:

{
  berthProfiles: profiles(serviceType: BERTH) {
    totalCount
  }
  youthProfiles: profiles(serviceType: YOUTH_MEMBERSHIP) {
    totalCount
  }
}

Since that has been possible before, that feature is still working.