HP-575 Make service input field in "addServiceConnection" mutation optional
Also marked the input field as deprecated.

When the `service` input field is provided, it overrides the Service
determined by authentication. The reason is that the requester can in
theory provide different `services` for different GraphQL
operations, within a single request, something like this:

mutation {
    add1: addServiceConnection(input: {
        serviceConnection: {
            service: {
                type: BERTH
            }
        }
    })

    add2: addServiceConnection(input: {
        serviceConnection: {
            service: {
                type: YOUTH_MEMBERSHIP
            }
        }
    })
}

Since that has been possible before, that feature is still working.

Put the Service into the request object in tests, like it will be
provided in the new system where the requester's Service is determined
by authentication.