Let test GraphQL client handle the request object construction
Many GraphQL tests had this kind of boilerplate:

    request = rf.post("/graphql")
    request.user = gql_client.user
    gql_client.execute(query, context=request)

That is now unnecessary as the test GraphQL client takes care of the
request object construction.

This commit changes the places where the User object is the only one
that was inserted into the request. Other tests add something else there
too.