HP-550 Handle encrypted search fields in NullsToEmptyStringsModel
If the model has an encrypted data field and a corresponding search
field, all data access should be handled via the search field and not
the data field. Hence the search field is included in the list of fields
to be checked, the data field is not.

The nullability needs to be checked from the data field, because the
search field doesn't reflect the data field's nullability but is always
True (as an implementation detail of django-searchable-encrypted-fields'
SearchField).