Avoid creating and immediately deleting empty addresses
The previous version of the code created each non-existing address model
of VerifiedPersonalInformation, and if the address was found to be
empty, it got immediately deleted afterwards. This caused two
unnecessary round trips to the database. Now the emptiness is checked
first and empty addresses aren't inserted into the database at all.