Avoid updating and immediately deleting empty addresses
The previous version of the code updated an address model of
VerifiedPersonalInformation, and if the address was found to be empty,
it got immediately deleted afterwards. This caused an unnecessary round
trip to the database. Now the emptiness is checked first and empty
addresses aren't unnecessarily updated into the database but just
deleted.

Removed the whole UpdateMixin. Perhaps it's not that useful...or if it
ever will be, it's easy to add again.