Remove not that useful test base class
The ProfileWithVerifiedPersonalInformationTestBase class contained only
some constants.

Everything within
test_gql_create_or_update_profile_with_verified_personal_information_mutation.py
was one big test class. Removed that test class and functions are now on
the top level of the file. Most of the changes in that file are because
of the reduced indentation.