HP-553 Ensure the context always has a service defined
Adds a new Graphene middleware called determine_service_middleware and enables
it in the settings. The middleware will determine service from the user
authentication or from the arguments and adds it into the context.

Additionally adds a decorator called login_and_service_required that can be
used to check that a user is authenticated and a service is really found.