HP-645 Make national_identification_number fields hash_key a callable
Before the hash_key value was a variable that would get evaluated when the field was declared creating a new migration for that field every time the hash_key changed. (e.g. in different environments)

Now the function name gets added to the migrations and won't change even if the SALT_NATIONAL_IDENTIFICATION_NUMBER setting changes.