Change UpdateProfileMutation to require an explicit service connection
Service connection wasn't checked on UpdateProfileMutation at all before the "Hp-553 Check for a service connection on read operations" patch. In that patch the service connection was checked but implicit connection was allowed. After that it was decided that explicit connection should be required to update a profile by a staff user.