Insert only the raw client id string to request
Previously the whole ServiceClientId model instance was inserted into
the request. But since the Service is in the request already as its own
attribute, it's enough to have the raw string representing the client id
there.

The pre-existing service_client_id attribute was not used anywhere so no
other code needed to be adapted.