Change the automatic null-to-empty-string field implementation
The previous implementation worked on the Model level, inspecting all
the fields of the Model. But this is a Field level functionality. Now
there's a little helper Field mixin that turns a null (i.e. None) into
an empty string just before saving to the database.

The mixin currently always turns a null into empty _string_, but it can
be easily extended in the future to support other types than just
strings, if needed.

Django wants to create a migration for this kind of Field type change,
but that migration has no effect on the database.