HP-843 Remove "old" JWT authentication system
The "old" system was based on django-heluser's ApiTokenAuthentication
class which is meant to be used with Django REST Framework kind of
setups. The newer system is based on django-heluser's
RequestJWTAuthentication and has no such dependency. From now on the
newer system is the only system.

The django-heluser's ApiTokenAuthentication class was the only part that
required the drf-oidc-auth library. That dependency is thus not needed
any more.