HP-876 Start building a Keycloak admin client
This client can be used to send requests to the Keycloak admin REST API.
Access to the API is granted with client credentials, a client id and
secret.

The client in this commit can fetch data for a user. It does access
token fetching every time a user's data is fetched and there's no error
handling, so this version isn't yet very usable. But it's a starting
point.