HP-876 Deny getting rid of a Profile's primary email
This applies to Profile modifications done via the GraphQL API. If a
Profile has a primary email, it can't be modified into a state where it
no longer has a primary email.

The actual implementation is quite short, but there are many entry
points in the GraphQL schema into this kind of modifications and all
those are now covered by tests. That's why there's proportionally quite
a lot of test code.