HP-593 Check amr claim value when accessing verified personal info
This new functionality is used to further secure the access to verified personal information.

"amr" (Authentication Methods Reference) claim is set by the OP to be the name of the authentication backend the user used to authenticate.

If the VERIFIED_PERSONAL_INFORMATION_ACCESS_AMR_LIST-setting is empty, any authentication method grants access.