HP-928 Change some Address model's fields' attributes
All these changes are packed into one commit because they make Django
want to create a migration file, even though nothing changes in the
database. This way one (no-op) migration file is enough.

It has been possible to insert empty strings into the Address fields via
GraphQL. Changed the fields' `blank` attributes to `True` to reflect the
reality correctly.

Changed the field types to `NullToEmptyCharField` so it automatically
turns a `null` value into an empty string.

Added a validator to the `country_code` field: only ISO 3166 alpha-2
type codes are accepted for this field. The field was already limited to
two characters, so alpha-2 type code is the only one that fits.