Make Profile serialization test cover the whole serialized data
The serialized profile data was only partially checked. The reason for
this was probably the fact that it wasn't reliable to check the
serialized data against a comparison object. The serialization process
produces lists of children objects. The order of those objects within
the list is not deterministic and can change from test run to another.

Solved this problem by introducing an unordered list type where the
order of members doesn't matter in the equality comparison. The
serialized data is translated to use this unordered list type before the
check against the comparison object. Now the serialized data is fully
checked.