Move migration test helper to a place where it can be used from everywhere
Previously the `migration_test_db` fixture didn't have a value. Its only
purpose was to setup the cleaning code that resets the database after
the test. Now the fixture has a value: it provides a function that is to
be called to execute the migration test. Renamed the fixture to
`execute_migration_test` accordingly.