Don't use Service in GraphQL tests that don't need one
The createOrUpdateUserProfile (and the older version of it,
createOrUpdateProfileWithVerifiedPersonalInformation, as well) is
supposed to work also when the caller doesn't identify any Service. This
is now verified in the tests. The Service is explicitly set to `None`,
because otherwise the test client's `execute` method puts a default
Service into the request.