Include migration files in flake8 runs
Fixed errors reported by flake8.

Add N806 to the ignored checks. That check enforces that variable names
in functions are lowercase. But it's quite customary in Django to write
code such as
    User = get_user_model()
which fetches a class into a variable that is named like a class.

Also added W503 to the ignored checks. Before this commit no ignored
checks were specified and the W503 is in the default ignored list
provided by flake8, so effectively it was already ignored previously.