Prototyping declarative data import
Provides a function that modifies the database data in a way that it
fulfills the provided declarative description. The prototype itself is
within a test file in order to speed up the prototyping. Naturally
that's not the correct place for a real implementation.

A leading design principle is to keep the declarative data description
as easy to understand as possible and natural for people to edit. If
that leads to more complex implementation in the importer, then so be
it.

- Requires natural keys, with exactly one field, for the models it
  handles.

- Supports foreign key relations as parent-child relations (see
  client_ids under Services).

- Supports many-to-many relations. The natural keys are used as the
  references (see allowed_data_fields under Services).

- Supports Django-parler translatable fields.

- Deletes all rows from the database that aren't specified in the
  declarative data (for those models that are touched by the declarative
  data). In a real implementation this shouldn't happen for all models
  but it should be controllable which models should be "purged" and
  which not.