Add system check for obsolete database tables
Extra tables can be left to the database for example if a previously
used app is removed from the project but its tables aren't cleaned from
the database. This check investigates what tables exist in the database,
what tables there should be according to the current configuration and
if there are any tables in the database that possibly shouldn't be there
any more.