Add system check for obsolete content types
Extra content types can be left to the database for example if a
previously used app is removed from the project but its content types
aren't cleaned from the database. This check investigates what content
types exist in the database, and if Django can provide a matching model
for them. If it can't, the content type is possibly some old remnant and
could be removed.

Bonus info: starting from Django 3.1 the `remove_stale_contenttypes`
management command has an `--include-stale-apps` flag which can be used
to remove these obsolete content types.
https://docs.djangoproject.com/en/3.1/ref/django-admin/#remove-stale-contenttypes