Fix pip-tools version in development dependencies
Pip-tools seems to change the format of the produced dependency file
quite often. It's better to freeze and document the used pip-tools
version in order to avoid confusion when pip-compile would otherwise
change the whole file on a simple dependency version update. The current
files were produced with pip-tools version 5.5.0 so fixed the version to
that now.