HP-1193 Add DEFAULT_AUTO_FIELD setting
Starting from version 3.2 Django will be migrating to another type for
auto generated primary key model fields. The type has been AutoField but
it shall change to BigAutoField in a future version. In order to avoid
needless migrations, the DEFAULT_AUTO_FIELD setting is set to the
historically used value of AutoField.

https://docs.djangoproject.com/en/3.2/releases/3.2/#customizing-type-of-auto-created-primary-keys