HP-790 By default order Profiles by `id` field
It's probably a good practice to have a deterministic and stable
ordering for any model. But it's especially required for Relay
pagination to work correctly. This commit adds a default ordering for
Profile model and changes the `profiles` GraphQL query test so that it
ensures that it really works.

Another test needed to be changed because it relied on data being read
from the database being in the same order as the data insertion order;
which off course isn't reliably true but usually works for small amounts
of data.