HP-790 Order Emails by `id` field, still putting primary email as first
It's probably a good practice to have a deterministic and stable
ordering for any model. But it's especially required for Relay
pagination to work correctly. This commit adds a default ordering for
Email model that works for any set of Emails. Previously any primary
email was put first, but the rest of the list was in unspecified order.
Now they are ordered by `id`.