HP-790 Order all contact items like Emails
Meaning by `id` field, but still putting the primary item as first.

It's probably a good practice to have a deterministic and stable
ordering for any model. But it's especially required for Relay
pagination to work correctly. This commit adds a default ordering for
the (abstract) Contact model, so that in addition to Emails, also
Addresses and Phones have the same ordering logic.