HP-790 Add test for AllowedDataFieldNode ordering
It's probably a good practice to have a deterministic and stable
ordering for any model. But it's especially required for Relay
pagination to work correctly.

The AllowedDataField model uses a SortableMixin from
django-admin-sortable (https://pypi.org/project/django-admin-sortable/).
That already defines an order for the model instances. A test was added
to check that the ordering is correct also in the GraphQL API.

The SortableMixin also forces the AllowedDataField model's `order` field
to the next available value upon creating a new instance. In order to be
able to control the `order` value in tests, a post_generation hook was
added to the AllowedDataFieldFactory which changes the `order` field
after the initial creation of the instance.