HP-1186 Add hel_translation directive and middleware
The new directive named "hel_translation" can be used on a field to
change the translation language of the field.

The directive requires one argument named "in". The argument's value
should be one of the choices in the new HelTranslationLanguageType
enum.

The HelTranslationLanguageType enum is automatically generated to
include every language in the "LANGUAGES" setting. The enum names are
uppercased versions of the language codes to keep consistent with how
enums are usually written in GraphQL.

Example GraphQL query:

    {
        hello_in_english: hello @hel_translation(in: EN)
        hello_in_finnish: hello @hel_translation(in: FI)
    }

The actual Django language (and Parler language when the fields root is
a TranslatableModel) change is done in the new
HelTranslationMiddleware which should be included in the "MIDDLEWARE"
list in the "GRAPHENE" settings.