HP-1234 Add new DjangoParlerObjectType type for Parler-translated models
The parler translated fields on a model are not added automatically
to the type when using DjangoObjectType because they are not
conventional Model fields.

The DjangoParlerObjectType is a subclass of DjangoObjectType which
adds a String field for all the translated fields in the model.
(If the field names are listed in "fields" parameter and not listed
in "exclude" parameter like in the default DjangoObjectType)

The added fields support one argument named "language" which can be used
to tell in which language the value should be in.

Example GraphQL query:

    {
        exampleModel {
            aField(language: EN)
        }
    }