Audit log reads in the admin
Adds a new AuditLogModelAdmin class for logging instance and inline
 instance reads in the admin.

The class overrides render_change_form which is used to render the
admin html. All of the querysets in the inline admin formsets are
looped and every instance in them are logged.

When using custom fields (like get_permanent_address in
VerifiedPersonalInformationAdminInline) we don't know from where the
value comes, so the audit logging must be handled manually.

Model instances in the list views are not logged.