Create primary key for UUIDModel lazily
If a model field (primary key or not) has a default value specified,
that field gets a value when an instance of that model is created in
memory. It doesn't make a difference whether that model later gets saved
into the database or not, the default value has already been set.

That behaviour is different from an auto incrementing primary key field.
An auto incrementing primary key field only gets a value after the model
has been saved. That's because it's the database that creates the value
so it can't be known before the save.

This commit changes the `UUIDModel` to behave like a model with an auto
incrementing primary key. This allows to check `instance.pk` to see
whether this instance exists in the database or not.

The previous behaviour also caused an extra audit log entry when
creating a new Profile. The creation of the Profile instance (in memory)
emitted an errorneous "READ" audit log entry. Saving the instance to the
database then emitted a "CREATE" entry, as expected. This weird
behaviour was reflected in the tests, but now that test only expects to
find the "CREATE" audit log entry.