Change Profile update audit log test to use GraphQL
The previous test just updated data directly in the database and assumed
that to create audit log entries. While that assumption seemed to work
thus far, it might not be correct. Audit log entries are required to be
formed only when accessing external APIs of the service, e.g. the
GraphQL API.