HP-1409 Change audit logging to "buffered" type - again
The previous implementation produced the audit log entry immediately
when audit logging was requested. Producing the audit log entry could
require a roundtrip to the database to fetch a Profile's User. When many
Profiles are handled during an HTTP request, this potentially leads to
many database queries during that request as well, to the point that
it's detrimental to performance.

In this commit the audit logging is divided into two steps. First audit
log data is gathered. Then audit log entries are produced according to
that data. This process is tied to an HTTP request: data is gathered
during the HTTP request handling and audit log entries are produced in
the end of the request. When the audit log entries are produced in the
end of an HTTP request, optimized database queries can be used to fetch
any additional needed data.

This change is very similar to a previous attempt in commit
8f96d1573acb4339b5b067aba959f309b9065e49. That commit was later reverted
(277ad95e5f18d0c40d214584e5a3877672d16acd) because of some issues with
that implementation.