HP-1409 Produce only one audit log entry of each type
The audit logging data gathering phase could produce multiple instances
of the same action for the same Profile part. In order to avoid
unnecessary duplicates, now only the first of these instances is logged,
the rest are discarded.

Changed also the tests to check that there are no multiple entries for
the same thing.