HP-1081 Don't send profile_updated signal within a transaction
The profile_updated signal currently causes changed Profile data to be
sent to Keycloak. As user data in Keycloak changes, it's possible that
Keycloak will call back to open-city-profile with some changed Profile
data. If a transaction is already open in open-city-profile for the same
Profile, that blocks the call coming from Keycloak accessing that
Profile's data and a dead lock occurs.

This commit changes the behaviour in open-city-profile a little.
Previously if sending the Profile data to Keycloak failed (an exception
got raised) the Profile changes in open-city-profile were also rolled
back. This is not the case any more.