HP-1639 Deactivate getting connections for a service
The GraphQL schema's ServiceNode type has a field called
serviceconnectionSet. That field has returned the connections related to
a service. Since it's also possible to get the service of a service
connection, this leads to a possibility to perform a recursive query:
get service connections and their services and their service connections
and their services etc.

A more serious effect was that there was no access control for receiving
the connections of a service. It was possible to receive _all_
connections related to a service. Although the service connections don't
contain any sensitive information, this was still wrong.

It was judged that there is no real need to get all connections of a
service. Hence the serviceconnectionSet fields now always returns an
empty result.