HP-1525 Read GDPR delete errors from the service
The GDPR API schema states that the service could return an "errors" key
when they are denying the profile deletion.

Change ServiceConnection.delete_gdpr_data to return results instead
of only raising exception if the deletion was not successful.

The results are inspected in the DeleteMyProfileMutation and exception
is raised there if there were errors.