HP-1429 Disable django-guardian's anonymous user support
Django-guardian supports giving Django's permissions to anonymous users.
This feature has an issue when a field is added or removed in the User
model: an exception gets thrown after the migration is run [1].

Open-city-profile doesn't currently use this feature. It can be disabled
by setting the `ANONYMOUS_USER_NAME` setting to `None` [2].

[1] https://github.com/django-guardian/django-guardian/issues/770
[2] https://django-guardian.readthedocs.io/en/stable/configuration.html#anonymous-user-name