HP-1050 Report an error when service lacks GDPR query scope
Previously the service without a query scope was just skipped. But
that's not good. When the user downloads their complete data, they
expect to receive all the data. If the configuration is lacking and not
all data from services can be queried, the user must be informed that
they didn't receive all their data.