HP-1787 Replace profile_updated signal with direct Keycloak integration
The `profile_updated` signal's idea was to decouple profile updates via
GraphQL from the Keycloak integration. The signal was sent after Profile
data had been updated, all the way to the database.

However, it turned out that Keycloak data needs to be updated before
Profile data gets updated. This will be easier to implement, when the
connection to the Keycloak integration is direct.

Additionally, GraphQL module was already tightly coupled with the
Keycloak integration, as it uses the `delete_profile_from_keycloak`
function directly.

No call to `send_profile_changes_to_keycloak` was added to the
ClaimProfileMutation because it is certain that the user doing the
profile claiming does not exist in Keycloak.