Add tests to check that only own related objects can be updated/removed
The email, phone and address objects of a profile are identified by a
global id when updated or removed. It's important that any given id
points to an object that belongs to the profile that is being
manipulated. The implementation already ensures this but now there are
also tests for it.